Add files via upload
Made the following changes to the code after the meeting:    1) Removed the custom QUBO function, lets only use PyQUBO    2) Added a picture from the QROSS paper that Paweł Gora pointed out.    3)Added comments pointing out where the equations are implemented in the code and where the relaxation parameter is located.